# docs/api-contracts/admin-module-openapi.yaml
...
openapi: 3.0.3
info:
  title: Admin Module API
  description: |
    Admin Module — управляющий модуль системы Artstore.

    **Admin Module не является auth-сервером.** Аутентификация и выдача JWT
    полностью делегированы внешнему Identity Provider (Keycloak). Admin Module
    получает JWT от API Gateway, проверяет claims (роли, scopes) и принимает
    решения по авторизации для своих endpoints.

    ## Аутентификация

    Все запросы (кроме Health) проходят через API Gateway, который:
    - Проверяет подпись JWT через JWKS Keycloak
    - Проверяет expiration
    - Форвардит claims в заголовках к Admin Module

    ### Admin Users (H2M)
    Аутентификация через Keycloak OIDC (Authorization Code + PKCE).
    JWT содержит `sub`, `realm_access.roles`, `groups`.

    Роли определяются маппингом Keycloak groups → Artstore roles:
    - `artstore-admins` → `admin` (полный доступ)
    - `artstore-viewers` → `readonly` (только чтение)

    Admin Module может **дополнить** (но не понизить) роль через
    локальные role overrides.

    ### Service Accounts (M2M)
    OAuth 2.0 Client Credentials flow через Keycloak.
    JWT содержит `client_id`, `scope`.

    Scopes:
    - `files:read` — чтение метаданных и скачивание файлов
    - `files:write` — загрузка, обновление и удаление файлов
    - `storage:read` — чтение информации о Storage Elements
    - `storage:write` — управление SE (sync, mode transition)
    - `admin:read` — чтение административных данных
...

    ## Shared PostgreSQL
...

...

...

...
servers:
...
paths:
...
  /api/v1/admin-users:
    get:
      tags: [admin-users]
      summary: Список пользователей
      description: |
        Возвращает пагинированный список пользователей.
        Данные берутся из Keycloak (через Admin API) и дополняются
        локальными role overrides из Admin Module.

        Доступно ролям `admin` и `readonly`.
      operationId: listAdminUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin-users/{id}:
    get:
      tags: [admin-users]
      summary: Получить пользователя
      description: |
        Возвращает данные пользователя по Keycloak user ID.
        Включает данные из Keycloak + локальные дополнения ролей.

        Доступно ролям `admin` и `readonly`.
      operationId: getAdminUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [admin-users]
      summary: Обновить локальные дополнения пользователя
      description: |
        Обновление локальных дополнений пользователя (role override).
        Может только **повысить** роль, не понизить.

        Создание и удаление пользователей, сброс паролей —
        выполняются в Keycloak.

        Доступно только роли `admin`.
      operationId: updateAdminUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Пользователь обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [admin-users]
      summary: Удалить локальные дополнения пользователя
      description: |
        Удаляет локальные дополнения (role override) для пользователя.
        Пользователь продолжит существовать в Keycloak с ролью из IdP.

        Доступно только роли `admin`.
      operationId: deleteAdminUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: Локальные дополнения удалены
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/admin-users/{id}/role-override:
    post:
      tags: [admin-users]
      summary: Установить/изменить локальное дополнение роли
      description: |
        Устанавливает или изменяет локальное дополнение роли для пользователя.

        **Правила:**
        - Можно только **повысить** роль (readonly → admin), не понизить
        - Итоговая роль = max(роль из IdP, локальное дополнение)
        - Если у пользователя в IdP уже `admin`, override `readonly`
          не имеет эффекта (но сохраняется для логирования)

        Доступно только роли `admin`.
      operationId: setRoleOverride
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleOverrideRequest"
      responses:
        "200":
          description: Role override установлен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Пользователь не найден в Keycloak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Service Accounts (6 endpoints)
  # =========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT от Keycloak. Валидация подписи — на API Gateway.
        Admin Module проверяет claims (роли, scopes).

        Для Admin Users: `realm_access.roles`, `groups`
        Для Service Accounts: `scope`

  # =========================================================================
  # Параметры
  # =========================================================================
  parameters:
    Limit:
      name: limit
      in: query
      description: Количество записей на странице
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка
      schema:
        type: integer
        minimum: 0
        default: 0
    UserId:
      name: id
      in: path
      required: true
      description: Keycloak user ID (UUID)
      schema:
        type: string
...
  responses:
    Unauthorized:
      description: Отсутствует или невалидный JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: "Требуется аутентификация"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: "Недостаточно прав для выполнения операции"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: "Ресурс не найден"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: "Внутренняя ошибка сервера"

  # =========================================================================
  # Схемы данных
  # =========================================================================
  schemas:
...
    AdminUser:
      type: object
      description: Пользователь (данные из Keycloak + локальные дополнения)
      required:
        - id
        - username
        - idp_role
        - effective_role
        - created_at
      properties:
        id:
          type: string
          description: Keycloak user ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        username:
          type: string
          example: alice
        email:
          type: string
          format: email
          nullable: true
          example: alice@example.com
        first_name:
          type: string
          nullable: true
          example: Alice
        last_name:
          type: string
          nullable: true
          example: Johnson
        groups:
          type: array
          items:
            type: string
          description: Группы пользователя из IdP
          example: ["artstore-admins"]
        idp_role:
          type: string
          enum: [admin, readonly]
          description: Роль из IdP
          example: readonly
        role_override:
          type: string
          enum: [admin, readonly]
          nullable: true
          description: Локальное дополнение роли
          example: admin
        effective_role:
          type: string
          enum: [admin, readonly]
          description: Итоговая роль = max(idp_role, role_override)
          example: admin
        enabled:
          type: boolean
          description: Активен ли аккаунт в Keycloak
          example: true
        created_at:
          type: string
          format: date-time
          description: Дата создания в Keycloak
          example: "2026-02-21T10:00:00Z"
    AdminUserUpdate:
      type: object
      description: Обновление локальных дополнений пользователя
      properties:
        role_override:
          type: string
          enum: [admin, readonly]
          nullable: true
          description: |
            Локальное дополнение роли.
            Установите `null` для удаления override.
      minProperties: 1
    RoleOverrideRequest:
      type: object
      description: Установка локального дополнения роли
      required:
        - role
      properties:
        role:
          type: string
          enum: [admin, readonly]
          description: Роль для локального дополнения
          example: admin
    AdminUserListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
          example: 5
        limit:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0
        has_more:
          type: boolean
          example: false

    # -----------------------------------------------------------------------
    # Service Accounts
    # -----------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Стандартный формат ошибки (единый для всей системы Artstore)
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: |
                Машиночитаемый код ошибки:
                - `VALIDATION_ERROR` — некорректные входные данные
                - `NOT_FOUND` — ресурс не найден
                - `UNAUTHORIZED` — требуется аутентификация
                - `FORBIDDEN` — недостаточно прав
                - `CONFLICT` — конфликт (дублирующийся ресурс)
                - `SE_UNAVAILABLE` — SE недоступен
                - `IDP_UNAVAILABLE` — Keycloak недоступен
                - `INTERNAL_ERROR` — внутренняя ошибка
              example: NOT_FOUND
            message:
              type: string
              description: Человекочитаемое описание
              example: Ресурс не найден